refactor(component): presentational component와 container  component로 분리한다
