+++ src/App.tsx
@@ -1,6 +1,5 @@
 import React from "react";
 import { css } from "@emotion/react";
-import Card from "./components/atoms/Card";
 import CardNumberFieldset from "./components/molecules/CardNumberFieldset";
 import AppProvider from "./components/Provider";
 import CardOwnerNameFieldset from "./components/molecules/CardOwnerNameFieldset";
@@ -9,6 +8,7 @@ import CVCFieldset from "./components/molecules/CVCFieldset";
 import CardPasswordFieldset from "./components/molecules/CardPasswordFieldset";
 import Navigation from "./components/molecules/Navigation";
 import ConfirmButtonContainer from "./components/atoms/confirm-button/ConfirmButtonContainer";
+import CardContainer from "./components/atoms/card/CardContainer";
 
 const style = css({
   width: '375px',
@@ -28,7 +28,7 @@ function App() {
       <Navigation />
       <AppProvider>
         <div css={cardWrapper}>
-          <Card />
+          <CardContainer />
         </div>
         <CardNumberFieldset />
         <ExpiredPeriodFieldset />
